SharedGameController edit
