--- src/main/java/com/scriptopia/demo/controller/TagDefController.java
@@ -1,29 +1,0 @@
-package com.scriptopia.demo.controller;
-
-import com.scriptopia.demo.dto.TagDef.TagDefCreateRequest;
-import com.scriptopia.demo.dto.TagDef.TagDefDeleteRequest;
-import com.scriptopia.demo.service.TagDefService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.core.Authentication;
-import org.springframework.web.bind.annotation.*;
-
-@RestController("/shared-games/tags")
-@RequiredArgsConstructor
-public class TagDefController {
-    private final TagDefService tagDefService;
-
-    @PreAuthorize("hasAnyAuthority('ADMIN')")
-    @PostMapping
-    public ResponseEntity<?> addTag(@RequestBody TagDefCreateRequest req) {
-
-        return tagDefService.addTagName(req);
-    }
-
-    @PreAuthorize("hasAnyAuthority('ADMIN')")
-    @DeleteMapping
-    public ResponseEntity<?> removeTag(@RequestBody TagDefDeleteRequest req) {
-        return tagDefService.removeTagName(req);
-    }
-}
